Testing now in place
Problems with race conditions in testing made things a pain, so it was
a struggle to make sure things are lined up properly. Now checks that we
can start up services and their fsms, and the state is reflected properly.
One potential issue is that after exiting the service, it's hard to
guarantee that we are asking the fsm for the state to be DOWN after the
monitor detects the crash.